+++ Source/TimeWarp.State/Components/TimeWarpStateComponent.cs
@@ -19,9 +19,6 @@ public partial class TimeWarpStateComponent : ComponentBase, IDisposable, ITimeW
 
   private static readonly ConcurrentDictionary<string, int> InstanceCounts = new();
   private static readonly ConcurrentDictionary<string, int> RenderCounts = new();
-  private static readonly ConcurrentDictionary<Type, string> ConfiguredRenderModeCache = new();
-  private static readonly ConcurrentDictionary<Type, bool> TypeRenderAttributeCache = new();
-
   /// <summary>
   ///   A generated unique Id based on the Class name and number of times they have been created
   /// </summary>
@@ -39,42 +36,7 @@ public partial class TimeWarpStateComponent : ComponentBase, IDisposable, ITimeW
   /// Set this to true if something in the component has changed that requires a re-render.
   /// </summary>
   protected bool NeedsRerender;
-  private bool HasRendered;
-  private bool UsesRenderMode;
   
-  protected string ConfiguredRenderMode =>
-    ConfiguredRenderModeCache.GetOrAdd(this.GetType(), type =>
-    {
-      // Use reflection to get all attributes on the current component type.
-      object[] attributes = type.GetCustomAttributes(true);
-
-      foreach (object attribute in attributes)
-      {
-        // Check if the type name of the attribute contains the expected name.
-        Type attributeType = attribute.GetType();
-        if (attributeType.Name.Contains("PrivateComponentRenderModeAttribute"))
-        {
-          // Try to get the 'Mode' property value of the attribute.
-          PropertyInfo? modeProperty = attributeType.GetProperty("Mode");
-          if (modeProperty != null)
-          {
-            // Use dynamic to bypass compile-time type checking
-            dynamic? modeValue = modeProperty.GetValue(attribute);
-            // Return the type name of the Mode property's value.
-            return modeValue == null ? "None" : modeValue.GetType().Name;
-          }
-        }
-      }
-
-      // If no matching attribute is found, return a default identifier.
-      return "None";// Adjust as needed for your default case.
-    });
-
-  /// <summary>
-  ///   Indicates if the component is being prerendered.
-  /// </summary>
-  protected bool IsPreRendering => GetCurrentRenderMode() == State.CurrentRenderMode.PreRendering;
-  protected string CurrentRenderMode => GetCurrentRenderMode().ToString();
   public int RenderCount => RenderCounts.GetValueOrDefault(Id, 0);
   
   public TimeWarpStateComponent()
@@ -172,21 +134,6 @@ public partial class TimeWarpStateComponent : ComponentBase, IDisposable, ITimeW
   
   private T? GetPreviousState<T>() => Store.GetPreviousState<T>();
   
-  private CurrentRenderMode GetCurrentRenderMode()
-  {
-    UsesRenderMode = true;
-    if (OperatingSystem.IsBrowser()) return State.CurrentRenderMode.Wasm;
-
-    if (HasRendered) return State.CurrentRenderMode.Server;
-
-    bool hasRenderAttribute = TypeRenderAttributeCache.GetOrAdd(this.GetType(), type =>
-      type.GetCustomAttributes(true)
-        .Any(attr => attr.GetType().Name.Contains("PrivateComponentRenderModeAttribute")));
-
-    return hasRenderAttribute ? State.CurrentRenderMode.PreRendering : State.CurrentRenderMode.Static;
-  }
-
-
   /// <summary>
   ///   Exposes StateHasChanged
   /// </summary>
@@ -196,12 +143,4 @@ public partial class TimeWarpStateComponent : ComponentBase, IDisposable, ITimeW
   {
     RenderCounts.AddOrUpdate(Id, 1, (_, count) => count + 1);
   }
-}
-
-public enum CurrentRenderMode
-{
-  Static,
-  PreRendering,
-  Server,
-  Wasm,
 }
